Changed clock source to external crystal rather than LSI, timing on the RTC is now accurate!
I actually grabbed a 8 MHz crystal off of the debug/programmer header from another nucleo board and plopped it down on the X3 solder pad on the L0 nucleo board. This 8 MHz was a complete guess because I know nothing about crystals but but some HAL magic I guess it worked!